passing interop tests with  /JsonWebSignature2020 instead of JsonWebSignature2020
