+++ packages/json-web-signature-2020/src/__fixtures__/contexts/credentials-v2.json
@@ -1,59 +1,62 @@
 {
-  "@context": {
-    "@version": 1.1,
-    "@protected": true,
+  "@context": [
+    {
+      "@version": 1.1
+    },
+    "https://www.w3.org/2018/credentials/v1",
+    {
+      "JsonWebSignature2020": {
+        "@id": "https://w3id.org/security#JsonWebSignature2020",
+        "@context": {
+          "@version": 1.1,
+          "@protected": true,
 
-    "JsonWebSignature2020": {
-      "@id": "https://w3id.org/security#JsonWebSignature2020",
-      "@context": {
-        "@version": 1.1,
-        "@protected": true,
+          "id": "@id",
+          "type": "@type",
 
-        "id": "@id",
-        "type": "@type",
+          "sec": "https://w3id.org/security#",
+          "xsd": "http://www.w3.org/2001/XMLSchema#",
 
-        "sec": "https://w3id.org/security#",
-        "xsd": "http://www.w3.org/2001/XMLSchema#",
+          "challenge": "sec:challenge",
+          "created": {
+            "@id": "http://purl.org/dc/terms/created",
+            "@type": "xsd:dateTime"
+          },
+          "domain": "sec:domain",
+          "expires": { "@id": "sec:expiration", "@type": "xsd:dateTime" },
+          "jws": "sec:jws",
+          "nonce": "sec:nonce",
+          "proofPurpose": {
+            "@id": "sec:proofPurpose",
+            "@type": "@vocab",
+            "@context": {
+              "@version": 1.1,
+              "@protected": true,
 
-        "challenge": "sec:challenge",
-        "created": {
-          "@id": "http://purl.org/dc/terms/created",
-          "@type": "xsd:dateTime"
-        },
-        "domain": "sec:domain",
-        "expires": { "@id": "sec:expiration", "@type": "xsd:dateTime" },
-        "jws": "sec:jws",
-        "nonce": "sec:nonce",
-        "proofPurpose": {
-          "@id": "sec:proofPurpose",
-          "@type": "@vocab",
-          "@context": {
-            "@version": 1.1,
-            "@protected": true,
+              "id": "@id",
+              "type": "@type",
 
-            "id": "@id",
-            "type": "@type",
+              "sec": "https://w3id.org/security#",
 
-            "sec": "https://w3id.org/security#",
-
-            "assertionMethod": {
-              "@id": "sec:assertionMethod",
-              "@type": "@id",
-              "@container": "@set"
-            },
-            "authentication": {
-              "@id": "sec:authenticationMethod",
-              "@type": "@id",
-              "@container": "@set"
+              "assertionMethod": {
+                "@id": "sec:assertionMethod",
+                "@type": "@id",
+                "@container": "@set"
+              },
+              "authentication": {
+                "@id": "sec:authenticationMethod",
+                "@type": "@id",
+                "@container": "@set"
+              }
             }
+          },
+          "proofValue": "sec:proofValue",
+          "verificationMethod": {
+            "@id": "sec:verificationMethod",
+            "@type": "@id"
           }
-        },
-        "proofValue": "sec:proofValue",
-        "verificationMethod": {
-          "@id": "sec:verificationMethod",
-          "@type": "@id"
         }
       }
     }
-  }
+  ]
 }
