+++ packages/json-web-signature-2020/src/__fixtures__/documentLoader.ts
@@ -10,21 +10,20 @@ const getJson = async (url: string) =>
     method: 'get',
   }).then((data: any) => data.json());
 
-// 'credentials-v1': require('./credentials-v1.json'),
-// 'credentials-v2': require('./credentials-v2.json'),
-// 'security-v2': require('./security-v2.json'),
-// 'security-v3': require('./security-v3.json'),
 const localOverrides: any = {
-  // eslint-disable-next-line
   [unlockedDID.id]: unlockedDID,
   'https://w3id.org/did/v0.11': require('./contexts/did-v0.11.json'),
+  // note the absence of:
+  // https://w3id.org/security/v1
+  // https://w3id.org/security/v2
+  // they are never called !
+  // leave commented out... does not exit...
+  // 'https://w3id.org/security/v3': require('./contexts/security-v3.json'),
   'https://www.w3.org/2018/credentials/v1': require('./contexts/credentials-v1.json'),
+  // leave commented out... does not exit...
+  // 'https://www.w3.org/2018/credentials/v2': require('./contexts/credentials-v2.json'),
   'https://www.w3.org/2018/credentials/examples/v1': require('./contexts/examples-v1.json'),
   'https://www.w3.org/ns/odrl.jsonld': require('./contexts/odrl.json'),
-  // 'https://www.w3.org/2018/credentials/v1': contexts['credentials-v1'],
-  // 'https://www.w3.org/2018/credentials/v2': contexts['credentials-v2'],
-  // 'https://w3id.org/security/v2': contexts['security-v2'],
-  // 'https://w3id.org/security/v3': contexts['security-v3'],
 };
 
 export const documentLoader = async (url: string) => {
@@ -51,6 +50,7 @@ export const documentLoader = async (url: string) => {
       documentUrl: url, // this is the actual context URL after redirects
     };
   }
+  // leave commented out... debugging with remote contexts is not recommended.
   // try {
   //   // console.log('downloading...', url);
   //   const res = await jsonld.documentLoader(url);
